Copied across modified scripts and executables from a testing repository
The window factor can be specified for CIJ in 'script_setup_analysis.py'.

The window factors for X and G are hardwired in 'x_pklX.py' and 'x_G.py', respectively.

The '--strong_signal' option (hence the output file name) needs to be changed
in 'x_postanalyse.py' manually.

These setup files and executables, as they are now, are set for analysis with a hanning window.

	modified:   pipeline/executables/x_G.py
	modified:   pipeline/executables/x_S.py
	modified:   pipeline/executables/x_analyse.py
	modified:   pipeline/executables/x_analyse_test.py
	modified:   pipeline/executables/x_pklTStoPSD.py
	modified:   pipeline/executables/x_pklX.py
	modified:   pipeline/executables/x_postanalyse.py
	modified:   pipeline/executables/x_stdP.py
	new file:   pipeline/executables/x_temp_optimals.py
	modified:   pipeline/scripts/script_setup_analysis.py
	modified:   pipeline/scripts/script_setup_post_analysis.py