'window' options added/used to X, G and S computation.
The option 'window' is added to 'x_G.py' and 'x_S.py'.
This option indicates whether no window or a hanning window is used
on the time-series.  The value of the window is then computed for each sample
of the time-series, and then the corresponding windowing factor is computed
for the estimators of X, G and S. ( For X, this window option is parsed to AS.XXX(AS.TimeSeries())
which calculates the windowing for the cross-spectra data cIJ. )