Enabled 'signal limit' option for stdP and stdPP.
In 'script_setup_postanalyse.py', the dictionaries 'stdP' and 'stdPP' now
have key 'signal limit' that indicates if the sigmas will be computed for
the weak signal limit, or for the strong signal limit, or for both.

In 'x_postanalyse.py', 'signal limit' key is parsed as an option of 'x_stdPP.py'
and 'x_stdP.py'.  For the output file name, if it is for weak-signal limit,
it has no additional tag.  If it is for the strong-signal limit, it has '_strong'.

In 'x_temp_optimals.py', the names of the files for the sigmas have been edited
according to the above.