Made 'echo done' changes to automate entire pipeline for testing.
Previous 'end-of-job' indicators based on writing to an initialised
empty list have been removed.  There were conficts in writing to the
same file at the same time perhaps.

Now, each PBS job submitted will have at the end of its output file
the line 'done'. A file with the same name as the output file is created
just before the submission of the job.  After the job is submitted, this file
( or the acutal output file, when the job is finished ) is read, and its last line
checked for 'done'.  This is done in a 'while' loop.  Once all the output files
end in 'done' for a given stage, the next part of the pipeline
(which requires results from the previous part) can be executed.

	modified:   pipeline/executables/x_P_to_TS.py
	modified:   pipeline/executables/x_analyse.py
	modified:   pipeline/executables/x_merge_ts_files.py
	modified:   pipeline/executables/x_merge_tss.py
	modified:   pipeline/executables/x_pklX.py
	modified:   pipeline/executables/x_postanalyse.py
	modified:   pipeline/executables/x_simulate_noise_for_days.py
	modified:   pipeline/executables/x_simulate_signal.py
	modified:   pipeline/executables/x_simulate_stationary_noise.py