Removed 'print done' statements from submitted executables.
Since in the master test script, 'done' is checked for at the
end of the output files of submitted jobs, additonal 'done' are
removed from those submitted executables. This prevents the test
script from thinking that a particular stage of the pipeline is over,
when only one day of the year has been completed.