Made some prelimary changes for recovering realisation
The algorithm for simulating for days is to set the seed at the very beginning.
If there are Nts time-series on each day, and, for each time-series,
the covariances at Nf frequencies are required, then for each day,
the number of random numbers drawn is Nf*2*Nts.  The factor of 2 is
because the covariance is usually complex, so needs two numbers,
 one for the real part and one for the imaginary part.

So, after setting the seed, ( D - 1 )*Nf*2Nts random numbers have
to be drawn before simulating for day D. This is implemented in
x_simulate_noise_for_days.py

	modified:   non_pipeline/executables/x_simulate_noise_for_days.py
	new file:   non_pipeline/executables/x_simulate_noise_for_days.py~