Seed-parsing changes made to the files below.
numpy.random.seed() only has effect on np.random.standard_nornaml() when
they are in the same session.  (e.g. if np.random.seed is applied before
a call to a function in which np.random.standard_normal is nested, then
np.random.seed has no effect on the call np.random.standard_normal in this case.)
Therefore, np.random.seed and np.random.standard_normal are both placed in
myUsefuls.get_noise_freq_domain_CovarMatrix(). And the scripts and executables
are used to parse to this function the SEED and the number of draws to discard.

The seed should either be 'random', to indicate that a random seed be chosen,
or an integer in quotes, like '45', to indicate that a seed of 45 be set.

	modified:   modules/myUsefuls.py
	modified:   non_pipeline/executables/x_simulate_noise_for_days.py
	modified:   non_pipeline/executables/x_simulate_stationary_noise.py
	modified:   non_pipeline/scripts/script_simulate_noise_for_days.py