Created myUsefuls.divide_days_in_batches_reverse_pairup()
This function divides a sequence of days(positive integers)
by half. Reverse one half. Pair up the halves. Then divide the pairs
into groups/batches of the same size.

This is useful for signal simulation, where the latter the day is
in the year, the longer its time-series takes to simulate, because
there are more random numbers to draw and discarde.  In order to submit jobs
for all the days such that they run for roughly the same time, the latter days
are paired up with earlier days of the year to balance out the simulation time.