Scripts made to wait 0.1s between processing each batch
This is really because there has been a puzzling problem whereby no output files
are written to disk for submittd jobs which have completed successfully.
This behaviour was observed for a set of simpler test scripts I wrote.
It was also observed however that things work fine /gpfs2/mapp_directory and in
\gpfs1/JR0XXXX.  The cause of this difference is unknown.

However, Jorge made the suggestion to makde a slight delay between
job submissions, which seems to have made the test scripts work.  So now,
the same changes are made to the following simulation scripts.

	modified:   non_pipeline/scripts/script_simulate_AETnoise_from_arbitrary_SpH_for_days.py
	modified:   non_pipeline/scripts/script_simulate_noise_for_days.py
	modified:   non_pipeline/scripts/script_simulate_noise_for_days_stitching.py