Added print statements to show stitched segments not matching
	modified:   non_pipeline/executables/x_simulate_noise_for_days_stitching.py

Segments are windowed and joined in pairs first, and the outputs from all
pairs are then concatenated.
Print statements have been added, which show that the last element of the output
of a pair does NOT match up with the first element of the output of the next pair.
For example, the last element of time-series (1|2) is not equal to the first
element of the time-series (3|4).
This might indicate a problem, as the aim of windowing and joining the segments
IS to avoid the discontinuity that happens when they are simply concatenated.