Print statements showing that tails are correctly passed on
	modified:   non_pipeline/executables/x_simulate_noise_for_days_stitching.py

When a pair of segments are windowed and joined together, a tail
(a time-series with duration half that of a segment) remains.
This is needs to be passed on to the next pair, when its segments are
windowed and joined.
These print statements print out the first 3 elements of the tails,
BEFORE and AFTER the window-and-join of a pair.  They show that before
the first pair, the tails are NONE.  Afterwards, the tail AFTER the window-and-join
is the same as the tail BEFORE the next window-and-join.