Set testsim.make_arbitrary_tdiORF_SpHs() to use orfAA's at DAY001
We would now like to use the same coloured signal in the two outputs
for the normalisation test.  The colour to be used will mimic that
of a realistic LISA's AA on DAY001.  Therefore, the frequency-dependent
g00 that has been used so far needs to be set to g00(f) of orfAA.

	modified:   modules/testsim.py