Added qsub's depend:afterok functionality to merging scripts and executables
Same idea as before, but as the merging of noise and signal depend on
both their simulations to finish, it needs to be made sure that
the ALL the jobids from them are put in the same text file ('afterok file') first.
This is up to the the user of the merging routines to do this.

	modified:   pipeline/executables/x_merge_tss.py
	modified:   pipeline/scripts/script_setup_merge_tss.py