Some parts of postanalysis pipeline linked with analysis pipeline with qsub -W depend=afterok:
This was done for sections do_P, do_PP, do_stdP and do_stdPP.
For do_P and do_PP, in script_setup_postanalysis.py, there is a new parameter for setting
whether to wait for jobs x_pklX and x_G from the analysis.
For do_stdP and do_stdPP, in script_setup_postanalysis.py, there is also a new parameter for setting
whether to wait for job x_G, in the case of weak-signal limit, or to wait for
job x_G and x_S from the analysis, in the case of strong-signal limit, or not wait at all.

	modified:   pipeline/executables/x_postanalyse.py
	modified:   pipeline/scripts/script_setup_post_analysis.py