sp_linalg.eig() changed to sp_linalg.eigh() in myUsefuls.get_noise_freq_domain_CovarMatrix()
This is a significant correction.  scipy.linalg.eig() does not does not
output a matrix of orthonormal eigenvectors, which is required by the simulation
method. This is necessarily done using
scipy.linalg.eigh(), which applies to Hermitian matrices.  This is caused by
confusion with Matlab, whose eig() automatically works on Hermitian matrices
and produce orthnormal eigenvectors.

After this change, the normalisation test carried out for the case of
different powerlaw (1/f) coloured time-series simulated from a single
covariance matrix is passed.  The reconstructed PSDs and CSDs,
averaged over 365 and 25 realisations, are shown to
agree with the analytical ones better than before.  The time and
realisation-averaged empirical mean of the data still fluctuates by as much
as before for freq-domain simulations, much larger than direct randn() sim
(for the white noise case).  The same averages for the variance and the covariance
are appear more or less the same as before.  However, for the 1-sigma test,
the optimal estimates more or less fluctuate evenly about the true value, and 70%
fall within 1 theoretical sigma.  This is the first time that a simulation
with coloured time-series (generated from a single covariance matrix having
non-zero off-diagonal elements) has worked.

	modified:   modules/myUsefuls.py