Added and edited some functions in PostProcess.py for projecting a SkyMap object using GNUPLOT
These new functions plot the pixels of a SkyMap object on a grid of ecliptic longitudes and latitudes.

plot_P.py is changed to used these functions
script_setup_post_analysis.py is changed so that all plotting tasks have the option of waiting
for the computation of the thing they're plotting to finish.

	modified:   modules/PostProcess.py
	modified:   pipeline/executables/plot_P.py
	modified:   pipeline/scripts/script_setup_post_analysis.py